+++ ElvUI_MerathilisUI/Core/Changelog/6.55.lua
@@ -1,8 +1,15 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[655] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "07.03.2025",
+	FIXES = {
+		"[Simple AddOn Manager]: Fix an error with the latest code changes",
+	},
 	NEW = {},
-	IMPROVEMENTS = {},
+	IMPROVEMENTS = {
+		"[World Map]: Updated MapData to 11.1",
+		"[World Map]: Reenable Event Tracker",
+		"[AutoButtons]: Add more openables,",
+		"[Misc]: Small code cleanups",
+	},
 }