+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ### Changes ###
 
-  * [SpecBar]: Fixed an issue which prevent so switch the Specialization- #104
+  * [Simple AddOn Manager]: Fix an error with the latest code changes
+  * [World Map]: Updated MapData to 11.1
+  * [World Map]: Reenable Event Tracker
+  * [AutoButtons]: Add more openables
+  * [Misc]: Small code cleanups